requirements.txt
Django
Faker
gunicorn
Pillow
python-dateutil
pytz
six
text-unidecode
whitenoise
celery==4.4.2
redis==3.2.1
djangorestframework
